Fix python cyclic imports for CodeQL
