+++ python/copilot/__init__.py
@@ -37,15 +37,11 @@ from .client import (
     LogLevel,
     ModelBilling,
     ModelCapabilities,
-    ModelCapabilitiesOverride,
     ModelInfo,
     ModelLimits,
-    ModelLimitsOverride,
     ModelPolicy,
     ModelSupports,
-    ModelSupportsOverride,
     ModelVisionLimits,
-    ModelVisionLimitsOverride,
     PingResponse,
     RemoteSessionMode,
     RuntimeConnection,
@@ -97,7 +93,10 @@ from .session import (
     MCPHTTPServerConfig,
     MCPServerConfig,
     MCPStdioServerConfig,
-    MemoryConfiguration,
+    ModelCapabilitiesOverride,
+    ModelLimitsOverride,
+    ModelSupportsOverride,
+    ModelVisionLimitsOverride,
     PermissionHandler,
     PermissionNoResult,
     PermissionRequestResult,
@@ -202,7 +201,6 @@ __all__ = [
     "MCPHTTPServerConfig",
     "MCPServerConfig",
     "MCPStdioServerConfig",
-    "MemoryConfiguration",
     "ModelBilling",
     "ModelCapabilities",
     "ModelCapabilitiesOverride",
