+++ python/copilot/session.py
@@ -83,7 +83,6 @@ logger = logging.getLogger(__name__)
 
 
 if TYPE_CHECKING:
-    from .client import ModelCapabilitiesOverride
     from .session_fs_provider import SessionFsProvider
 
 # Re-export SessionEvent under an alias used internally
@@ -92,6 +91,66 @@ SessionEventTypeAlias = SessionEvent
 # ============================================================================
 # Reasoning Effort
 # ============================================================================
+
+@dataclass
+class ModelVisionLimitsOverride:
+    supported_media_types: list[str] | None = None
+    max_prompt_images: int | None = None
+    max_prompt_image_size: int | None = None
+
+
+@dataclass
+class ModelLimitsOverride:
+    max_prompt_tokens: int | None = None
+    max_output_tokens: int | None = None
+    max_context_window_tokens: int | None = None
+    vision: ModelVisionLimitsOverride | None = None
+
+
+@dataclass
+class ModelSupportsOverride:
+    vision: bool | None = None
+    reasoning_effort: bool | None = None
+
+
+@dataclass
+class ModelCapabilitiesOverride:
+    supports: ModelSupportsOverride | None = None
+    limits: ModelLimitsOverride | None = None
+
+
+def _capabilities_to_dict(caps: ModelCapabilitiesOverride) -> dict:
+    result: dict = {}
+    if caps.supports is not None:
+        s: dict = {}
+        if caps.supports.vision is not None:
+            s["vision"] = caps.supports.vision
+        if caps.supports.reasoning_effort is not None:
+            s["reasoningEffort"] = caps.supports.reasoning_effort
+        if s:
+            result["supports"] = s
+    if caps.limits is not None:
+        lim: dict = {}
+        if caps.limits.max_prompt_tokens is not None:
+            lim["max_prompt_tokens"] = caps.limits.max_prompt_tokens
+        if caps.limits.max_output_tokens is not None:
+            lim["max_output_tokens"] = caps.limits.max_output_tokens
+        if caps.limits.max_context_window_tokens is not None:
+            lim["max_context_window_tokens"] = caps.limits.max_context_window_tokens
+        if caps.limits.vision is not None:
+            v: dict = {}
+            if caps.limits.vision.supported_media_types is not None:
+                v["supported_media_types"] = caps.limits.vision.supported_media_types
+            if caps.limits.vision.max_prompt_images is not None:
+                v["max_prompt_images"] = caps.limits.vision.max_prompt_images
+            if caps.limits.vision.max_prompt_image_size is not None:
+                v["max_prompt_image_size"] = caps.limits.vision.max_prompt_image_size
+            if v:
+                lim["vision"] = v
+        if lim:
+            result["limits"] = lim
+    return result
+
 
 ReasoningEffort = Literal["low", "medium", "high", "xhigh"]
 ReasoningSummary = Literal["none", "concise", "detailed"]
@@ -2453,8 +2512,6 @@ class CopilotSession:
         """
         rpc_caps = None
         if model_capabilities is not None:
-            from .client import _capabilities_to_dict
-
             rpc_caps = _RpcModelCapabilitiesOverride.from_dict(
                 _capabilities_to_dict(model_capabilities)
             )
